# Add compile time selection for copying data with specialised paths