--- a/src/openrct2/core/MemoryStream.h
+++ b/src/openrct2/core/MemoryStream.h
@@ -58,7 +58,51 @@ public:
     void Seek(int64_t offset, int32_t origin) override;
 
     void Read(void* buffer, uint64_t length) override;
+    void Read1(void* buffer) override;
+    void Read2(void* buffer) override;
+    void Read4(void* buffer) override;
+    void Read8(void* buffer) override;
+    void Read16(void* buffer) override;
+
+    template<size_t N> void Read(void* buffer)
+    {
+        uint64_t position = GetPosition();
+        if (position + N > _dataSize)
+        {
+            throw IOException("Attempted to read past end of stream.");
+        }
+
+        std::memcpy(buffer, _position, N);
+        _position = (void*)((uintptr_t)_position + N);
+    }
+
     void Write(const void* buffer, uint64_t length) override;
+    void Write1(const void* buffer) override;
+    void Write2(const void* buffer) override;
+    void Write4(const void* buffer) override;
+    void Write8(const void* buffer) override;
+    void Write16(const void* buffer) override;
+
+    template<size_t N> void Write(const void* buffer)
+    {
+        uint64_t position = GetPosition();
+        uint64_t nextPosition = position + N;
+        if (nextPosition > _dataCapacity)
+        {
+            if (_access & MEMORY_ACCESS::OWNER)
+            {
+                EnsureCapacity((size_t)nextPosition);
+            }
+            else
+            {
+                throw IOException("Attempted to write past end of stream.");
+            }
+        }
+
+        std::memcpy(_position, buffer, N);
+        _position = (void*)((uintptr_t)_position + N);
+        _dataSize = std::max<size_t>(_dataSize, (size_t)nextPosition);
+    }
 
     uint64_t TryRead(void* buffer, uint64_t length) override;