This improves the performance of IStream/MemoryStream most where the effect is most prominent with the recently introduced gamestate snapshots system.

openrct2 host parks/bpb.sv6:
```
In-Game FPS bpb.sv6 at 1080p Software (HW):
  Average Before: 39
  Average After: 56

GameStateSnapshots::Capture:
  Average Before: 0.003934 secs
  Average After: 0.002075 secs
```

This should also affect other areas where those functions are used, there are 280~ references to Read/ReadValue and 110~ references to Write/WriteValue