@Gymnasiast 2019-05-30
Result of doing `openrct2 benchgfx sc0.sc4 100`:
```

Without:
Rendering 100 times with drawing engine Software took 1.74 seconds. = 57 FPS

With:
Rendering 100 times with drawing engine Software took 1.63 seconds. = 61 FPS
```

I love how you managed to squeeze out more FPS, even though OpenRCT2 went through a lot of optimisations already.